requirements.txt
hrv==0.2.10
matplotlib==3.5.2
numpy==1.23.1
py_ecg_detectors==1.3.2
